+++ include/melon/algorithm/dijkstra.hpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <cassert>
 #include <concepts>
+#include <cstddef>
 #include <optional>
 #include <ranges>
 #include <type_traits>