+++ include/melon/algorithm/network_voronoi.hpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <cassert>
 #include <concepts>
+#include <cstddef>
 #include <ranges>
 #include <type_traits>
 #include <utility>