+++ include/melon/container/static_forward_digraph.hpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <cassert>
 #include <concepts>
+#include <cstddef>
 #include <numeric>
 #include <ranges>
 #include <span>